[REFACTOR/#42] 기록 화면 제작
