+++ app/src/main/java/com/example/momolabfe/ui/consult/ConsultFragment.kt
@@ -22,10 +22,6 @@ class ConsultFragment : Fragment() {
         return binding.root
     }
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-    }
-
     override fun onResume() {
         super.onResume()
         val bottomNav = activity?.findViewById<BottomNavigationView>(R.id.main_bnv)
