+++ app/src/main/java/com/example/momolabfe/ui/record/RecordListFragment.kt
@@ -24,11 +24,9 @@ class RecordListFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+
+        // 바텀 내비게이션 숨기기
+        activity?.findViewById<BottomNavigationView>(R.id.main_bnv)?.visibility = View.GONE
     }
 
-    override fun onResume() {
-        super.onResume()
-        val bottomNav = activity?.findViewById<BottomNavigationView>(R.id.main_bnv)
-        bottomNav?.visibility = View.VISIBLE
-    }
 }
