+++ app/src/main/java/com/example/momolabfe/ui/record/RecordOcrFragment.kt
@@ -6,20 +6,19 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.momolabfe.R
-import com.example.momolabfe.databinding.FragmentRecordWrite01Binding
+import com.example.momolabfe.databinding.FragmentRecordOcrBinding
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
-class RecordWrite01Fragment : Fragment() {
+class RecordOcrFragment : Fragment() {
 
-    private var _binding: FragmentRecordWrite01Binding? = null
+    private var _binding: FragmentRecordOcrBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentRecordWrite01Binding.inflate(inflater, container, false)
-
+        _binding = FragmentRecordOcrBinding.inflate(inflater, container, false)
         return binding.root
     }
 
@@ -29,9 +28,20 @@ class RecordWrite01Fragment : Fragment() {
         // 바텀 내비게이션 숨기기
         activity?.findViewById<BottomNavigationView>(R.id.main_bnv)?.visibility = View.GONE
 
-        binding.nextBtn.setOnClickListener {
+        binding.closeIv.setOnClickListener {
+            parentFragmentManager.popBackStack()
+        }
+
+        binding.cameraCv.setOnClickListener {
             parentFragmentManager.beginTransaction()
-                .replace(R.id.main_frm, RecordWrite02Fragment())
+                .replace(R.id.main_frm, RecordWrite01Fragment()) // 카메라
+                .addToBackStack(null)
+                .commit()
+        }
+
+        binding.albumCv.setOnClickListener {
+            parentFragmentManager.beginTransaction()
+                .replace(R.id.main_frm, RecordWrite01Fragment()) // 앨범
                 .addToBackStack(null)
                 .commit()
         }
