+++ app/src/main/java/com/example/momolabfe/ui/record/RecordWrite02Fragment.kt
@@ -1,7 +1,5 @@
 package com.example.momolabfe.ui.record
 
-import android.net.Uri
-import android.os.Build
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -9,15 +7,15 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import com.example.momolabfe.R
-import com.example.momolabfe.databinding.FragmentRecord02Binding
+import com.example.momolabfe.databinding.FragmentRecordWrite02Binding
 import com.example.momolabfe.ui.record.viewModel.RecordViewModel
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class Record02Fragment : Fragment() {
+class RecordWrite02Fragment : Fragment() {
 
-    private var _binding: FragmentRecord02Binding? = null
+    private var _binding: FragmentRecordWrite02Binding? = null
     private val binding get() = _binding!!
 
     private val viewModel: RecordViewModel by activityViewModels()
@@ -32,7 +30,7 @@ class Record02Fragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentRecord02Binding.inflate(inflater, container, false)
+        _binding = FragmentRecordWrite02Binding.inflate(inflater, container, false)
         return binding.root
     }
 
@@ -41,22 +39,7 @@ class Record02Fragment : Fragment() {
 
         // 바텀 내비게이션 숨기기
         activity?.findViewById<BottomNavigationView>(R.id.main_bnv)?.visibility = View.GONE
-
-        // 넘겨준 이미지 Uri로 OCR 호출
-        val imageUri: Uri? = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            arguments?.getParcelable("image_uri", Uri::class.java)
-        } else {
-            @Suppress("DEPRECATION")
-            arguments?.getParcelable("image_uri")
-        }
-
-        if (imageUri != null) {
-            viewModel.recordByOcr(imageUri)
-        }
-
     }
-
-
 
     override fun onDestroyView() {
         super.onDestroyView()
