+++ app/src/main/res/drawable/ic_album_sv.xml
@@ -1,0 +1,21 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:aapt="http://schemas.android.com/aapt"
+    android:width="40dp"
+    android:height="40dp"
+    android:viewportWidth="25"
+    android:viewportHeight="25">
+  <path
+      android:pathData="M20.833,6.25C21.386,6.25 21.916,6.469 22.306,6.86C22.697,7.251 22.917,7.781 22.917,8.333V20.139C22.917,20.691 22.697,21.221 22.306,21.612C21.916,22.002 21.386,22.222 20.833,22.222H7.292C6.739,22.222 6.209,22.002 5.819,21.612C5.428,21.221 5.208,20.691 5.208,20.139V8.333C5.208,7.781 5.428,7.251 5.819,6.86C6.209,6.469 6.739,6.25 7.292,6.25H20.833ZM12.221,14.525L7.433,19.313L7.365,19.375L7.292,19.431V20.14H20.833V19.431L20.76,19.376L20.692,19.313L18.358,16.98L17.991,17.349L18.048,17.406C18.235,17.594 18.343,17.845 18.351,18.11C18.359,18.374 18.266,18.632 18.091,18.83C17.916,19.029 17.672,19.153 17.409,19.178C17.145,19.204 16.882,19.128 16.673,18.966L16.575,18.879L12.221,14.525ZM17.708,3.125C18.234,3.125 18.74,3.323 19.125,3.681C19.511,4.038 19.747,4.528 19.786,5.052L19.792,5.208H5.208C4.953,5.208 4.707,5.302 4.516,5.472C4.326,5.641 4.204,5.875 4.174,6.128L4.167,6.25V18.75C3.641,18.75 3.135,18.552 2.749,18.194C2.364,17.837 2.128,17.347 2.089,16.823L2.083,16.667V6.25C2.083,5.453 2.388,4.686 2.935,4.106C3.481,3.526 4.229,3.177 5.025,3.13L5.208,3.125H17.708ZM20.833,8.333H7.292V16.508L11.3,12.5C11.421,12.379 11.564,12.283 11.722,12.218C11.88,12.152 12.05,12.118 12.221,12.118C12.392,12.118 12.561,12.152 12.719,12.218C12.877,12.283 13.021,12.379 13.142,12.5L16.518,15.875L17.438,14.955C17.559,14.834 17.703,14.738 17.861,14.673C18.019,14.607 18.188,14.574 18.359,14.574C18.53,14.574 18.7,14.607 18.858,14.673C19.016,14.738 19.159,14.834 19.28,14.955L20.833,16.508V8.333ZM17.188,9.896C17.393,9.896 17.596,9.936 17.785,10.015C17.975,10.093 18.147,10.208 18.292,10.354C18.437,10.499 18.552,10.671 18.631,10.86C18.71,11.05 18.75,11.253 18.75,11.458C18.75,11.663 18.71,11.867 18.631,12.056C18.552,12.246 18.437,12.418 18.292,12.563C18.147,12.708 17.975,12.823 17.785,12.902C17.596,12.98 17.393,13.021 17.188,13.021C16.773,13.021 16.376,12.856 16.083,12.563C15.79,12.27 15.625,11.873 15.625,11.458C15.625,11.044 15.79,10.646 16.083,10.354C16.376,10.061 16.773,9.896 17.188,9.896Z">
+    <aapt:attr name="android:fillColor">
+      <gradient 
+          android:startX="12.5"
+          android:startY="3.125"
+          android:endX="12.5"
+          android:endY="22.222"
+          android:type="linear">
+        <item android:offset="0" android:color="#FFB549FB"/>
+        <item android:offset="1" android:color="#FFEC3CA9"/>
+      </gradient>
+    </aapt:attr>
+  </path>
+</vector>
