+++ app/src/main/res/drawable/ic_bulb_sv.xml
@@ -1,0 +1,30 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="25dp"
+    android:height="25dp"
+    android:viewportWidth="25"
+    android:viewportHeight="25">
+  <path
+      android:pathData="M10.938,22.656C10.938,23.516 11.641,24.219 12.5,24.219C13.359,24.219 14.063,23.516 14.063,22.656H10.938Z"
+      android:fillColor="#616466"/>
+  <path
+      android:pathData="M9.727,18.75H15.234C15.547,17.07 16.602,15.43 17.695,13.711C18.984,11.719 20.313,9.648 20.313,7.656C20.313,3.867 16.797,0.781 12.5,0.781C8.203,0.781 4.688,3.867 4.688,7.656C4.688,9.648 6.016,11.719 7.266,13.711C8.359,15.43 9.414,17.07 9.727,18.75Z"
+      android:fillColor="#FFCE31"/>
+  <path
+      android:pathData="M10.313,13.125C10.351,13.359 10.43,13.594 10.469,13.828C10.586,14.258 10.664,14.648 10.781,15.078C11.133,16.563 11.445,17.813 11.719,18.789H11.953C11.758,17.813 11.484,16.523 11.133,15.039C11.055,14.648 10.938,14.219 10.859,13.789C10.82,13.594 10.781,13.359 10.703,13.164C11.016,13.086 11.758,12.852 12.5,12.031C13.242,12.852 13.984,13.086 14.297,13.164C14.258,13.398 14.219,13.594 14.141,13.789C14.063,14.219 13.945,14.648 13.867,15.039C13.516,16.523 13.242,17.813 13.047,18.789H13.281C13.516,17.813 13.867,16.563 14.219,15.078C14.297,14.688 14.414,14.258 14.531,13.828C14.57,13.594 14.648,13.359 14.688,13.125C15,13.086 15.273,13.008 15.469,12.813C15.703,12.578 15.82,12.305 15.742,11.992C15.703,11.836 15.625,11.641 15.391,11.602C15.273,11.563 15.195,11.563 15.078,11.602C14.961,11.641 14.883,11.719 14.844,11.758C14.648,11.992 14.57,12.227 14.492,12.461L14.375,12.813C14.375,12.813 14.063,12.852 13.32,12.344C12.93,12.07 12.813,11.875 12.695,11.758C12.813,11.602 12.93,11.484 13.047,11.289C13.086,11.172 13.125,11.094 13.164,10.938C13.164,10.82 13.164,10.664 13.086,10.547C13.008,10.43 12.93,10.313 12.734,10.234C12.656,10.195 12.578,10.195 12.5,10.195C12.422,10.195 12.344,10.195 12.266,10.234C12.109,10.313 11.992,10.43 11.914,10.547C11.836,10.664 11.836,10.82 11.836,10.938C11.836,11.055 11.875,11.172 11.953,11.289C12.031,11.484 12.148,11.641 12.305,11.758C12.188,11.875 12.07,12.07 11.68,12.344C10.976,12.813 10.625,12.813 10.625,12.813L10.508,12.461C10.43,12.227 10.351,11.992 10.156,11.758C10.117,11.719 10.039,11.641 9.922,11.602C9.805,11.563 9.687,11.563 9.609,11.602C9.375,11.68 9.297,11.875 9.258,11.992C9.219,12.305 9.336,12.578 9.531,12.813C9.766,13.008 10.039,13.086 10.313,13.125ZM14.844,12.539C14.922,12.344 15,12.109 15.156,11.992C15.195,11.953 15.234,11.953 15.234,11.953H15.313C15.351,11.953 15.391,11.992 15.43,12.109C15.469,12.266 15.391,12.5 15.234,12.617C15.117,12.695 14.961,12.773 14.805,12.813C14.805,12.695 14.805,12.617 14.844,12.539ZM12.266,10.664C12.344,10.586 12.422,10.508 12.5,10.508C12.578,10.508 12.656,10.547 12.734,10.664C12.813,10.781 12.773,10.938 12.695,11.094C12.656,11.211 12.578,11.328 12.5,11.406C12.422,11.289 12.344,11.211 12.305,11.094C12.226,10.977 12.188,10.82 12.266,10.664ZM9.609,12.031C9.609,11.953 9.648,11.914 9.727,11.875H9.805C9.805,11.875 9.844,11.914 9.883,11.914C10,12.031 10.117,12.266 10.195,12.461C10.234,12.539 10.234,12.617 10.273,12.695C10.117,12.656 9.961,12.617 9.844,12.5C9.687,12.422 9.609,12.227 9.609,12.031Z"
+      android:fillColor="#C79127"/>
+  <path
+      android:pathData="M9.727,19.531H15.313V20.234H9.727V19.531ZM10.117,20.938H14.922V21.641H10.117V20.938Z"
+      android:fillColor="#94989B"/>
+  <path
+      android:pathData="M10.117,20.234H14.922V20.938H10.117V20.234Z"
+      android:fillColor="#616466"/>
+  <path
+      android:pathData="M15.313,19.531L10.117,20.938V21.68L15.313,20.234V19.531ZM10.508,22.383H14.531V23.086H10.508V22.383Z"
+      android:fillColor="#94989B"/>
+  <path
+      android:pathData="M10.508,21.68H14.531V22.383H10.508V21.68Z"
+      android:fillColor="#616466"/>
+  <path
+      android:pathData="M14.922,20.938L10.508,22.383V23.125L14.922,21.68V20.938Z"
+      android:fillColor="#94989B"/>
+</vector>
