+++ app/src/main/res/drawable/ic_home_sv.xml
@@ -1,6 +1,6 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
+    android:width="25dp"
+    android:height="25dp"
     android:viewportWidth="24"
     android:viewportHeight="24">
   <path
