+++ app/src/main/res/layout/fragment_consult.xml
@@ -98,10 +98,10 @@
                 android:layout_marginHorizontal="10dp"
                 android:backgroundTint="@color/white"
                 app:cardCornerRadius="10dp"
-                app:cardElevation="0dp"
+                app:cardElevation="1dp"
                 app:layout_constraintStart_toEndOf="@id/agent_iv"
                 app:layout_constraintTop_toTopOf="@id/agent_iv"
-                app:strokeWidth="1dp">
+                app:strokeWidth="0dp">
 
                 <androidx.constraintlayout.widget.ConstraintLayout
                     android:layout_width="match_parent"
