+++ app/src/main/res/layout/fragment_home.xml
@@ -64,7 +64,7 @@
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toBottomOf="@id/content_tv"
-                    app:strokeWidth="1dp">
+                    app:strokeWidth="0dp">
 
                     <androidx.constraintlayout.widget.ConstraintLayout
                         android:layout_width="match_parent"
@@ -112,7 +112,7 @@
 
                 <com.google.android.material.card.MaterialCardView
                     android:id="@+id/complete_cv"
-                    android:layout_width="180dp"
+                    android:layout_width="165dp"
                     android:layout_height="wrap_content"
                     android:layout_marginHorizontal="10dp"
                     android:layout_marginStart="20dp"
@@ -122,7 +122,7 @@
                     app:cardElevation="0dp"
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toBottomOf="@id/summary_tv"
-                    app:strokeWidth="1dp">
+                    app:strokeWidth="0dp">
 
                     <androidx.constraintlayout.widget.ConstraintLayout
                         android:layout_width="match_parent"
@@ -172,7 +172,7 @@
 
                 <com.google.android.material.card.MaterialCardView
                     android:id="@+id/uf_cv"
-                    android:layout_width="180dp"
+                    android:layout_width="165dp"
                     android:layout_height="wrap_content"
                     android:layout_marginHorizontal="10dp"
                     android:layout_marginStart="10dp"
@@ -182,7 +182,7 @@
                     app:cardElevation="0dp"
                     app:layout_constraintStart_toEndOf="@id/complete_cv"
                     app:layout_constraintTop_toBottomOf="@id/summary_tv"
-                    app:strokeWidth="1dp">
+                    app:strokeWidth="0dp">
 
                     <androidx.constraintlayout.widget.ConstraintLayout
                         android:layout_width="match_parent"
@@ -248,13 +248,13 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_marginTop="30dp"
-                    android:layout_marginEnd="10dp"
+                    android:layout_marginStart="220dp"
                     android:text="전체보기"
                     android:fontFamily="@font/noto_sans_kr_medium"
                     android:includeFontPadding="false"
                     android:textColor="@color/main_text"
                     android:textSize="16sp"
-                    app:layout_constraintEnd_toEndOf="@id/uf_cv"
+                    app:layout_constraintStart_toEndOf="@id/recent_record_tv"
                     app:layout_constraintTop_toBottomOf="@id/complete_cv"/>
 
                 <androidx.recyclerview.widget.RecyclerView
@@ -285,7 +285,7 @@
                     app:cardElevation="0dp"
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toBottomOf="@id/recent_record_rv"
-                    app:strokeWidth="1dp">
+                    app:strokeWidth="0dp">
 
                     <androidx.constraintlayout.widget.ConstraintLayout
                         android:layout_width="match_parent"
