+++ app/src/main/res/layout/fragment_statistics.xml
@@ -60,7 +60,7 @@
 
             <com.google.android.material.card.MaterialCardView
                 android:id="@+id/weekly_avg_weight_cv"
-                android:layout_width="180dp"
+                android:layout_width="165dp"
                 android:layout_height="wrap_content"
                 android:layout_marginHorizontal="10dp"
                 android:layout_marginStart="20dp"
@@ -70,7 +70,7 @@
                 app:cardElevation="0dp"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@id/title_container"
-                app:strokeWidth="1dp">
+                app:strokeWidth="0dp">
 
                 <androidx.constraintlayout.widget.ConstraintLayout
                     android:layout_width="match_parent"
@@ -117,7 +117,7 @@
 
             <com.google.android.material.card.MaterialCardView
                 android:id="@+id/weekly_avg_uf_cv"
-                android:layout_width="180dp"
+                android:layout_width="165dp"
                 android:layout_height="wrap_content"
                 android:layout_marginHorizontal="10dp"
                 android:layout_marginStart="10dp"
@@ -126,7 +126,7 @@
                 app:cardElevation="0dp"
                 app:layout_constraintStart_toEndOf="@id/weekly_avg_weight_cv"
                 app:layout_constraintTop_toTopOf="@id/weekly_avg_weight_cv"
-                app:strokeWidth="1dp">
+                app:strokeWidth="0dp">
 
                 <androidx.constraintlayout.widget.ConstraintLayout
                     android:layout_width="match_parent"
@@ -176,13 +176,12 @@
                 android:layout_width="180dp"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="20dp"
-                android:layout_marginHorizontal="10dp"
                 android:backgroundTint="@color/white"
                 app:cardCornerRadius="10dp"
-                app:cardElevation="0dp"
+                app:cardElevation="1dp"
                 app:layout_constraintStart_toStartOf="@id/weekly_avg_weight_cv"
                 app:layout_constraintTop_toBottomOf="@id/weekly_avg_weight_cv"
-                app:strokeWidth="1dp">
+                app:strokeWidth="0dp">
 
                 <androidx.constraintlayout.widget.ConstraintLayout
                     android:layout_width="match_parent"
