+++ app/src/main/res/menu/bottom_nav.xml
@@ -12,11 +12,6 @@
         android:title="기록"/>
 
     <item
-        android:id="@+id/fragment_record_list"
-        android:icon="@drawable/ic_record_list_sv"
-        android:title="목록"/>
-
-    <item
         android:id="@+id/fragment_consult"
         android:icon="@drawable/ic_consult_sv"
         android:title="상담"/>
